requirements.txt
opencv-python
numpy
mediapipe
streamlit
pygame
matplotlib
seaborn
pandas
pinecone-client
